Add nuget packages and update project file
This commit imports the necessary Microsoft.EntityFrameworkCore and Microsoft.Extensions.Logging.Abstractions packages and updates the project file accordingly. This change was made to incorporate the use of entity framework and logging abstractions to better manage tasks and processes in the existing project. The `DotFed.csproj` file was also updated to include references to additional packages for cryptography, diagnostics, in-memory database and RazorSlices that are essential for the functionality of the app.