Add initial version of DotFedWorker application
This commit introduces the base structure for the DotFedWorker application. The application will connect to a database, connect to a RabbitMQ, consume messages, and log the messages. Libraries needed for postgresql, RabbitMQ, and JSON parsing have been added. Also, a Dockerfile has been added for containerization. Further functions and refinements will be introduced in the subsequent commits.